first batch of simplifications

+++ mp_is_square.c
@@ -28,10 +28,10 @@ static const char rem_105[105] = {
 /* Store non-zero to ret if arg is square, and zero if not */
 mp_err mp_is_square(const mp_int *arg, bool *ret)
 {
-   mp_err        err;
-   mp_digit      c;
-   mp_int        t;
-   unsigned long r;
+   mp_err   err;
+   mp_digit c;
+   mp_int   t;
+   uint32_t r;
 
    /* Default to Non-square :) */
    *ret = false;
